requirements.txt
web3>=6.15
duckdb>=0.10
eth-abi>=5.0
python-dotenv>=1.0
pytest>=8.0
